Move .NET5 project's file into CSharpLatest
