+++ analyzers/its/sources/CSharpLatest/Net5/S2092.cs
