+++ analyzers/its/sources/CSharpLatest/Net5/S2115.cs
