+++ analyzers/its/sources/CSharpLatest/Net5/S3330.cs
