+++ analyzers/its/sources/CSharpLatest/Net5/S5122.cs
