+++ analyzers/its/sources/CSharpLatest/Net5/S5659.cs
