+++ analyzers/its/sources/CSharpLatest/Net5Test/Net5Test.csproj
