+++ analyzers/its/sources/CSharpLatest/Net5Test/S2699.cs
