+++ analyzers/its/sources/CSharpLatest/Net5Test/S3433_Xunit.cs
