+++ analyzers/its/sources/Net5/net5.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Net5", "Net5\Net5.csproj", "{F9326C7E-D6A2-47C1-9062-C9C63413C92A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Net5Test", "Net5Test\Net5Test.csproj", "{A5FA90FD-12AB-4E52-BE5D-601A54C2CFA6}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
